Toger#Partition bio modules
