+++ main.go
@@ -32,8 +32,8 @@ func route() {
 	http.HandleFunc("/statistics", model.AdvancedStatistic)
 	http.HandleFunc("/test/protocol", model.ProtocalHTTP)
 	http.HandleFunc("/tpl", model.HomeTemplate)
-	http.HandleFunc("/hg38/refgene/modes", bio.PayloadHg38Modes)
-	http.HandleFunc("/hg38/refgene/gene", bio.PayloadHg38RefGene)
+	http.HandleFunc("/hg38/refgene/modes", bio.Hg38RefgeneModes)
+	http.HandleFunc("/hg38/refgene/gene", bio.Hg38Refgene)
 	http.HandleFunc("/", model.Home)
 }
